Update ports.c fix lmao

+++ kernel/ports.c
@@ -17,7 +17,7 @@ uint8_t inb(uint16_t port) {
 }
 
 uint16_t inw(uint16_t port) {
-    uint8_t result;
+    uint16_t result;
     asm volatile ("inw %1, %0" : "=a" (result) : "Nd" (port));
     return result;
 }
